jpeg: Try to implement IDCT

+++ crates/graphics/image/src/jpeg/huffman_table.rs
@@ -82,8 +82,8 @@ struct HuffmanTableEntry {
 
 impl HuffmanTables {
     #[must_use]
-    pub fn get(&self, index: usize) -> Result<&HuffmanTable, Error> {
-        self.tables[index]
+    pub fn get(&self, index: u8) -> Result<&HuffmanTable, Error> {
+        self.tables[index as usize]
             .as_ref()
             .ok_or(Error::UndefinedHuffmanTable)
     }
